requirements.txt
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.2.0
requests>=2.28.0
python-dotenv>=0.19.0
geocoder>=1.38.1
joblib>=1.2.0
pytest>=7.4.0